+++ transpositiontable.go
@@ -7,8 +7,10 @@ import (
 type TranspositionEntry struct {
 	move      Move
 	depth     int8
-	score     int
-	entryType uint8 // technically this could just be 1-2 bits.
+	score     int16
+	entryType uint8
+	// technically this could just be 1-2 bits, we might want to give the
+	// other bytes to the score.
 }
 
 const TT_INITIAL_SIZE = 1048576
@@ -57,7 +59,7 @@ func (entry *TranspositionEntry) String() string {
 	return fmt.Sprintf("{score=%d, depth=%d, type=%s}", entry.score, entry.depth, entryTypeAsString)
 }
 
-func StoreTranspositionTable(boardState *BoardState, move Move, score int, entryType uint8, depth int8) {
+func StoreTranspositionTable(boardState *BoardState, move Move, score int16, entryType uint8, depth int8) {
 	// Try to avoid a new heap allocation if we already have something at this hash key.
 	var entry *TranspositionEntry
 	if boardState.transpositionTable[boardState.hashKey] != nil {
